Give access to production properties from Python.
Give access to a Well's production properties from Python. This is
just a draft currently, only a minimal API is added as a "proof of
concept".

A method get_well_status() is also added to demonstrate how to get
the status of well directly from the Schedule object. The current
way of getting the status of a well seems to have a problem.
It first calls get_well() on the Schedule object, and then calls
status() on the Well object. This works fine if the status is not
changed later from Python. However, if the status is changed by a
call to e.g. shut_well() on the Schedule object, the original Python
Well object does not reflect this change. I believe this is because
shut_well() calls updateWellStatus() in Schedule.cpp which creates a
new Opm::Well object by moving the old object into new one.