Read/modify schedule production properties from Python.
Make it possible to read/modify the Opm::Well and its production
properties that are stored a given index (time step) in the
Opm::Schedule object from Python. This is a draft, as only a minimal
API is added and the code currently does not work as expected.

The code compiles, but when modifying the oil rate limit from Python,
only the oil-rate limit at the given report step is modified. For
example, consider a production well PROD. If the oil rate for PROD
is modified from Python at report step "idx", and the original Schedule
(as initialized from the deck) did not specify any oil rate changes for
PROD on the following report steps (idx+1, idx+2,..), the rate for PROD
will switch back to the original one after the current step (idx+1).
The correct behavior would be if the modified oil rate for PROD stayed
that way until the next change well target change for PROD as specified
in the original Schedule.

Note that the current Pybind implementation for Opm::Schedule has
a method get_well() that can be used to get a reference to an
Opm::Well at a given time step index from Python. However this
well can be shared (it is a shared pointer) between different time
steps which makes it unreliable for doing updates to the Schedule
state. The same applies to a Well's WellProductionProperties
class variable. It is also a shared pointer that can be shared
between multiple wells.

This is the reason that this commit does not make use of
get_well() but instead works directly with the Opm::Schedule
object throught the Pybind11 methods get_well_production_target()
and set_well_production_target()