Read/modify schedule production properties from Python.
Make it possible to read/modify the Opm::Well and its production
properties that are stored a given index (time step) in the
Opm::Schedule object from Python. This is a draft, as only a minimal
API is added and the code currently does not work.

The code compiles, but when I try to run it by modifying the
oil rate limit from Python, and then run a simulation with
opm.BlackOilSimulator I get an exception after some time steps. The
error message is:

Opm::RestartIO::Helpers::WindowedArray<T>::Idx = long unsigned int]:
Assertion `(window < this->numWindows()) && "Window ID Out of Bounds"'
failed.

Note that the current Pybind implementation for Opm::Schedule has
a method get_well() that can be used to get a reference to an
Opm::Well at a given time step index from Python. However this
well can be shared (it is a shared pointer) between different time
steps which makes it unreliable for doing updates to the Schedule
state. The same applies to a Well's WellProductionProperties
class variable. It is also a shared pointer that can be shared
between multiple wells.

This is the reason that this commit does not make use of
get_well() but instead works directly with the Opm::Schedule
object throught the Pybind11 methods get_well_production_target()
and set_well_production_target()